Use a Map to hold the traversed path
I wasn't happy with the assumption that paths longer than the area of
the room should be discarded. However, since for part 1 I didn't bother
keeping track of the direction the guard was moving this was an easy way
to solve part 2. The code ran in ~10 seconds though which was acceptable.

This version uses a Map to hold the list of the directions the guard
was moving on each square. Therefore, loops can be detected as soon as
they occur.  Disappointingly slow at nearly 16 seconds.